+++ src/sqlfluff/rules/L036.py
@@ -104,8 +104,15 @@ class Rule_L036(BaseRule):
                 == select_target.pos_marker.working_line_no
             ):
                 # Find and delete any whitespace before the select target.
+                start_seg = select_targets_info.select_idx
+                # If any select modifier (e.g. distinct ) is present, start
+                # there rather than at the beginning.
+                modifier = segment.get_child("select_clause_modifier")
+                if modifier:
+                    start_seg = segment.segments.index(modifier)
+
                 ws_to_delete = segment.select_children(
-                    start_seg=segment.segments[select_targets_info.select_idx]
+                    start_seg=segment.segments[start_seg]
                     if not i
                     else select_targets_info.select_targets[i - 1],
                     select_if=lambda s: s.is_type("whitespace"),
@@ -133,12 +140,28 @@ class Rule_L036(BaseRule):
             < select_targets_info.first_new_line_idx
             < select_targets_info.first_select_target_idx
         ):
+            # Do we have a modifier?
+            modifier = select_clause.get_child("select_clause_modifier")
+
             # there is a newline between select and select target
             insert_buff = [
                 WhitespaceSegment(),
                 select_clause.segments[select_targets_info.first_select_target_idx],
                 NewlineSegment(),
             ]
+
+            # Also move any modifiers if present
+            if modifier:
+                # If it's already on the first line, ignore it.
+                if (
+                    select_clause.segments.index(modifier)
+                    < select_targets_info.first_new_line_idx
+                ):
+                    modifier = None
+                # Otherwise we need to move it too
+                else:
+                    insert_buff = [WhitespaceSegment(), modifier] + insert_buff
+
             fixes = [
                 # Replace "newline" with <<select_target>>, "newline".
                 LintFix(
@@ -152,6 +175,16 @@ class Rule_L036(BaseRule):
                     select_clause.segments[select_targets_info.first_select_target_idx],
                 ),
             ]
+
+            # Also delete the original modifier if present
+            if modifier:
+                fixes += [
+                    LintFix(
+                        "delete",
+                        modifier,
+                    ),
+                ]
+
             if (
                 select_targets_info.first_select_target_idx
                 - select_targets_info.first_new_line_idx
@@ -171,12 +204,12 @@ class Rule_L036(BaseRule):
                         ],
                     ),
                 )
-            if parent_stack and parent_stack[-1].type == "select_statement":
+            if parent_stack and parent_stack[-1].is_type("select_statement"):
                 select_stmt = parent_stack[-1]
                 select_clause_idx = select_stmt.segments.index(select_clause)
                 after_select_clause_idx = select_clause_idx + 1
                 if len(select_stmt.segments) > after_select_clause_idx:
-                    if select_stmt.segments[after_select_clause_idx].type == "newline":
+                    if select_stmt.segments[after_select_clause_idx].is_type("newline"):
                         # The select_clause is immediately followed by a
                         # newline. Delete the newline in order to avoid leaving
                         # behind an empty line after fix.